bp_xcm_bridge_hub_router

Trait XcmChannelStatusProvider

pub trait XcmChannelStatusProvider {
    // Required method
    fn is_congested(with: &Location) -> bool;
}

XCM channel status provider that may report whether it is congested or not.

By channel we mean the physical channel that is used to deliver messages of one of the bridge queues.

Required Methods

fn is_congested(with: &Location) -> bool

Returns true if the channel is currently congested.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.