Trait bp_xcm_bridge_hub_router::XcmChannelStatusProvider

pub trait XcmChannelStatusProvider {
    // Required method
    fn is_congested() -> bool;
}

XCM channel status provider that may report whether it is congested or not.

By channel we mean the physical channel that is used to deliver messages of one of the bridge queues.

Required Methods

fn is_congested() -> bool

Returns true if the channel is currently congested.

Object Safety

This trait is not object safe.